feat: very broken first version of map toggle
