+++ src/components/map/Map.svelte
@@ -6,7 +6,7 @@
 	import MiscMarkers from "$components/map/layers/MiscMarkers.svelte";
 	import POILabels from "$components/map/layers/POILabels.svelte";
 	import Grid from "$components/map/layers/Grid.svelte";
-	import LocationMarker from "$components/map/LocationMarker.svelte";
+	import WaypointMarker from "$components/map/WaypointMarker.svelte";
 	import MousePos from "$components/ui/widgets/MousePos.svelte";
 	import WaypointPos from "$components/ui/widgets/WaypointPos.svelte";
 	import LayerControl from "$components/ui/widgets/LayerControl.svelte";
@@ -14,16 +14,20 @@
 	import { map } from "$store";
 	import { isTouchDevice } from "$scripts/platform-check";
 
+	export let mapName;
 	export let mapData;
+	export let keys;
 
 	let mapContainer;
 
 	onMount(() => {
-		const bounds = [
-			[0, 0],
-			[mapData.height, mapData.width],
-		];
-		const image = L.imageOverlay(mapData.image, bounds);
+		setupMap(mapData);
+	});
+
+	function setupMap(mapData) {
+		if (!mapData || !mapContainer) return;
+
+		$map?.remove();
 
 		$map = L.map(mapContainer, {
 			crs: L.CRS.Simple,
@@ -32,6 +36,12 @@
 			zoomControl: false,
 		});
 
+		const bounds = [
+			[0, 0],
+			[mapData.height, mapData.width],
+		];
+
+		const image = L.imageOverlay(mapData.image, bounds);
 		image.addTo($map);
 
 		$map.fitBounds(bounds);
@@ -40,36 +50,32 @@
 			[mapData.height + 3000, mapData.width + 3000],
 		]);
 		$map.setZoom(-1);
-	});
+	}
+
+	$: setupMap(mapData);
 </script>
 
-{#if $map}
-	<CustomMarkers />
+{#key mapData}
+	{#if $map}
+		<CustomMarkers />
+		<KeyMarkers keys={keys.filter((key) => key.map === mapName)} />
+		{#each mapData?.locations?.misc ?? [] as misc}
+			<MiscMarkers {...misc} />
+		{/each}
 
-	{#if mapData?.locations?.keys}
-		<KeyMarkers keys={mapData.locations.keys} />
-	{/if}
-
-	{#each mapData?.locations?.misc as misc}
-		<MiscMarkers {...misc} />
-	{/each}
-
-	{#if mapData?.options?.pois}
 		<POILabels pois={mapData.options.pois} />
-	{/if}
-	{#if mapData?.options?.grid}
 		<Grid mapHeight={mapData.height} mapWidth={mapData.width} grid={mapData.options.grid} />
-	{/if}
 
-	<LocationMarker />
+		<WaypointMarker xMax={mapData.width} yMax={mapData.height} />
 
-	{#if !isTouchDevice()}
-		<MousePos />
+		{#if !isTouchDevice()}
+			<MousePos xMax={mapData.width} yMax={mapData.height} />
+		{/if}
+		<WaypointPos xMax={mapData.width} yMax={mapData.height} />
+		<ZoomControl />
+		<LayerControl />
 	{/if}
-	<WaypointPos />
-	<ZoomControl />
-	<LayerControl />
-{/if}
+{/key}
 <section bind:this={mapContainer} />
 
 <style>
