+++ src/data/ashika-island/data.js
@@ -1,4 +1,4 @@
-import image from "$assets/maps/al-mazrah.jpg";
+import image from "$assets/maps/ashika-island.jpg";
 import { pois } from "./locations/pois";
 import { keys } from "./locations/keys";
 import { deadDrops } from "./locations/dead-drops";
@@ -6,8 +6,8 @@ import { spawns } from "./locations/spawns";
 
 export const data = {
 	image,
-	width: 4150,
-	height: 4150,
+	width: 1210,
+	height: 1210,
 	options: {
 		pois,
 		grid: {
