+++ src/data/ashika-island/locations/dead-drops.js
@@ -1,0 +1,1 @@
+export const deadDrops = [];
