+++ src/data/ashika-island/locations/keys.js
@@ -1,0 +1,21 @@
+export const keys = [
+	{
+		title: "Tsuki Castle",
+		description: [""],
+		video: "pmlZu6GtZxg",
+		tags: ["fortress"],
+	},
+	{
+		title: "Tsuki Castle Weapons Locker",
+		description: [""],
+		video: "wP2LHs0LAQg",
+	},
+	{
+		title: "Beach Club Bathhouse",
+		description: [""],
+	},
+	{
+		title: "Main Harbor Control Key",
+		description: [""],
+	},
+];
