+++ src/data/ashika-island/locations/pois.js
@@ -1,0 +1,1 @@
+export const pois = [];
