+++ src/data/ashika-island/locations/spawns.js
@@ -1,0 +1,1 @@
+export const spawns = [];
