+++ src/data/keys.js
@@ -1,0 +1,1434 @@
+// This file gets generated automatically.
+// Please don't manually change anything in this file.
+// For key data changes refer to the keys.js file in each map directory. 
+
+export const keys = [
+    {
+        "id": "ahmed-grocery-store",
+        "map": "al-mazrah",
+        "title": "Ahmed Grocery Store",
+        "keyName": "Ahmed Grocery Store Office Key",
+        "location": {
+            "x": 3056,
+            "y": 2220
+        },
+        "description": [
+            "Easy to find, this is the office room inside of the grocery store."
+        ],
+        "video": "xfyHNVdPqQs"
+    },
+    {
+        "id": "airport-maintenance",
+        "map": "al-mazrah",
+        "title": "Airport Maintenance",
+        "location": {
+            "x": 2941,
+            "y": 2874
+        },
+        "description": [
+            "The Airport Maintenance is a room on the northeast side of the Al Malik Airport. There are multiple entrances on the ground floor on the inside and outside and another on top of the roof."
+        ],
+        "video": "zu5eO6MpaOE",
+        "keyName": "Airport Maintenance Key"
+    },
+    {
+        "id": "abf-antiquities-office",
+        "map": "al-mazrah",
+        "title": "A.B.F. Antiquities Office",
+        "location": {
+            "x": 2532,
+            "y": 3471
+        },
+        "description": [
+            "Inside the large hall in the middle of the fortress. Entrance is up the indoor balcony staircase on the west side of the hall."
+        ],
+        "video": "qTZby0PqlEw",
+        "keyName": "A.B.F. Antiquities Office Key"
+    },
+    {
+        "id": "al-bagra-barrack",
+        "map": "al-mazrah",
+        "title": "Al Bagra Barrack",
+        "location": {
+            "x": 2570,
+            "y": 3439
+        },
+        "description": [
+            "Access to both doors from the ground floor on the east side of the fortress. Each room can be opened separately (multiple key uses)."
+        ],
+        "video": "-eeyyCbW6j8",
+        "keyName": "Al Bagra Barrack Key"
+    },
+    {
+        "id": "al-bagra-munitions-storeroom",
+        "map": "al-mazrah",
+        "title": "Al Bagra Munitions Storeroom",
+        "location": {
+            "x": 2483,
+            "y": 3453
+        },
+        "description": [
+            "Inside a possible stronghold location, so a keycard might be needed. Bottom floor has two rooms with metal bars that can each be opened separately (multiple key uses)."
+        ],
+        "video": "mJCOg4vLYwE",
+        "keyName": "Al Bagra Munitions Storeroom Key"
+    },
+    {
+        "id": "al-bagra-officers-quarters",
+        "map": "al-mazrah",
+        "title": "Al Bagra Officers Quarters",
+        "location": {
+            "x": 2560,
+            "y": 3471
+        },
+        "description": [
+            "Located on the top floor of the eastern wing of the fortress. Includes two rooms that can each be opened separately (multiple key uses)."
+        ],
+        "video": "viPRIk2N3xw",
+        "keyName": "Al Bagra Officers Quarters Key"
+    },
+    {
+        "id": "al-bagra-underground",
+        "map": "al-mazrah",
+        "title": "Al Bagra Underground",
+        "location": {
+            "x": 2548,
+            "y": 3520
+        },
+        "description": [
+            "Required for Breaking and Entering, Black Mous tier 4.",
+            "Entrance from the road down on the south side of the fortress."
+        ],
+        "video": "qi_OlMLmp_A",
+        "tags": [
+            "fortress",
+            "mission"
+        ],
+        "keyName": "Al Bagra Underground Key"
+    },
+    {
+        "id": "algae-covered-toolbox",
+        "map": "al-mazrah",
+        "title": "Algae Covered Toolbox",
+        "location": {
+            "x": 2359,
+            "y": 1484
+        },
+        "description": [
+            "This toolbox can be found underwater between two bridges spanning over the canal in Al-Mazrah City. "
+        ],
+        "video": "dyKZnctOtvA",
+        "keyName": "Algae Covered Toolbox Key"
+    },
+    {
+        "id": "al-said-shopping-centre",
+        "map": "al-mazrah",
+        "title": "Al Sa'id Shopping Centre",
+        "location": {
+            "x": 1459,
+            "y": 2607
+        },
+        "description": [
+            "Multiple entrances can be found around the Shopping Centre, including in the back just underneath the pinned location. Be warned, this location usually contains a lot of enemies."
+        ],
+        "video": "YmLZEqov-iE",
+        "tags": [
+            "fortress"
+        ],
+        "keyName": "Al Sa'id Shopping Centre Key"
+    },
+    {
+        "id": "al-sharim-pass-elders-room",
+        "map": "al-mazrah",
+        "title": "Al Sharim Pass Elders Room",
+        "location": {
+            "x": 2599,
+            "y": 2488
+        },
+        "description": [
+            "The locked room is located on the roof of the building on top of the mountain at Al Sharim Pass."
+        ],
+        "video": "PoJimZ73oZA",
+        "keyName": "Al Sharim Pass Elders Room Key"
+    },
+    {
+        "id": "al-safwa-stone-block-office",
+        "map": "al-mazrah",
+        "title": "Al-Safwa Stone Block Office",
+        "location": {
+            "x": 902,
+            "y": 1834
+        },
+        "description": [
+            "The Stone Block Office is located to the southeast of Al-Safwa Quarry. There are doors on both sides of the small building."
+        ],
+        "video": "IjbsN4v810U",
+        "keyName": "Al-Safwa Stone Block Office Key"
+    },
+    {
+        "id": "art-museum",
+        "map": "al-mazrah",
+        "title": "Art Museum",
+        "location": {
+            "x": 2175,
+            "y": 1485
+        },
+        "description": [
+            "Multiple entrances around the large building to the west of Al-Mazrah City. As this location is a fortress it contains a lot of overzealous art fanatics."
+        ],
+        "video": "U0QfE_-KQJA",
+        "tags": [
+            "fortress"
+        ],
+        "keyName": "Art Museum Key"
+    },
+    {
+        "id": "bc-toolbox",
+        "map": "al-mazrah",
+        "title": "B.C. Toolbox",
+        "location": {
+            "x": 2836,
+            "y": 2854
+        },
+        "description": [
+            "In the north of the large airport building there is a room behind the Border Control desks. The toolbox is located in this room near the conveyor belts.",
+            "The key can be acquired by completing Tactical Extraction, Legion tier 3"
+        ],
+        "video": "ARSEpO4vBcE",
+        "keyName": "B.C. Toolbox Key"
+    },
+    {
+        "id": "bank-of-adal-top-story-office",
+        "map": "al-mazrah",
+        "title": "Bank of Adal Top Story Office",
+        "location": {
+            "x": 1281,
+            "y": 3273
+        },
+        "description": [
+            "Found on the top floor inside the bank in Sawah Village. The room is located in the southeast corner of the floor."
+        ],
+        "video": "ikdMfq1syeY",
+        "keyName": "Bank of Adal Top Story Office Key"
+    },
+    {
+        "id": "bank-of-adal-2nd-story-office",
+        "map": "al-mazrah",
+        "title": "Bank of Adal 2nd Story Office",
+        "location": {
+            "x": 1274,
+            "y": 3272
+        },
+        "description": [
+            "This locked door is located on the 2nd story of the Bank of Adal in Sawah Village. Going up the stairs you will find the door on your right."
+        ],
+        "video": "WkT4jwhecYU",
+        "keyName": "Bank of Adal 2nd Story Office Key"
+    },
+    {
+        "id": "canal-apartment-103",
+        "map": "al-mazrah",
+        "title": "Canal Apartment 103",
+        "location": {
+            "x": 2332,
+            "y": 1401
+        },
+        "description": [
+            "Access to the apartment building can be found from the side of the road. The room is located on the west side of the ground floor."
+        ],
+        "video": "at0B9x2A4w4",
+        "keyName": "Canal Apartment 103 Key"
+    },
+    {
+        "id": "captains-cache",
+        "map": "al-mazrah",
+        "title": "Captain's Cache",
+        "location": {
+            "x": 1822,
+            "y": 3566
+        },
+        "description": [
+            "Climbing up the shipwreck you will find a cabin. The cache is inside this cabin on the west side, just above the water that's flooded the cabin."
+        ],
+        "video": "hyUB1Aq3R1I",
+        "keyName": "Captain's Cache Key"
+    },
+    {
+        "id": "caretakers-house",
+        "map": "al-mazrah",
+        "title": "Caretaker's House",
+        "location": {
+            "x": 1619,
+            "y": 3115
+        },
+        "description": [
+            "The Caretakers House is located to the far south east of the Al Samman Cemetery."
+        ],
+        "video": "L1SIzWnabRU",
+        "keyName": "Caretaker's House Key"
+    },
+    {
+        "id": "caretakers-toolbox",
+        "map": "al-mazrah",
+        "title": "Caretaker's Toolbox",
+        "location": {
+            "x": 1588,
+            "y": 2994
+        },
+        "description": [
+            "This toolbox can be a bit difficult to find. It is located between the graves in the Al Samman Cemetery, about 5-6 rows in from the north wall."
+        ],
+        "video": "pgUwvHx8S90",
+        "keyName": "Caretaker's Toolbox Key"
+    },
+    {
+        "id": "cavern-boat-dock-shack",
+        "map": "al-mazrah",
+        "title": "Cavern Boat Dock Shack",
+        "location": {
+            "x": 1485,
+            "y": 2200
+        },
+        "description": [
+            "This small shack can be found next to the docks in the west side of the underground cave system. You can enter this cave by boat or through the various entrances around the mountain. Note that there are usually a lot of guards around, happy to defend their little pirate cove.",
+            "The key can be acquired by completing Big Spender, Legion tier 2"
+        ],
+        "video": "o3JWaK8UPIA",
+        "keyName": "Cavern Boat Dock Shack Key"
+    },
+    {
+        "id": "central-said-top-floor-apartment",
+        "map": "al-mazrah",
+        "title": "Central Sa'id Top Floor Apartment",
+        "location": {
+            "x": 1206,
+            "y": 2673
+        },
+        "description": [
+            "The apartment is on the top floor of the second building from the southwest in Sa'id City."
+        ],
+        "video": "LfLFurevylc",
+        "keyName": "Central Sa'id Top Floor Apartment Key"
+    },
+    {
+        "id": "central-zaya-meeting-room",
+        "map": "al-mazrah",
+        "title": "Central Zaya Meeting Room",
+        "location": {
+            "x": 1993,
+            "y": 2491
+        },
+        "description": [
+            "The building is located at the Zaya Observatory, it can spawn as a stronghold so a keycard might be needed. The Meeting Room is on the first floor right in front of the top of the staircase."
+        ],
+        "video": "FlR0jpT5uS8",
+        "keyName": "Central Zaya Meeting Room Key"
+    },
+    {
+        "id": "chemical-storage-warehouse",
+        "map": "al-mazrah",
+        "title": "Chemical Storage Warehouse",
+        "location": {
+            "x": 763,
+            "y": 1969
+        },
+        "description": [
+            "The West Al-Safwa Chemical Storage Warehouse is found to the far west of the map, just outside of Al-Safwa Quarry. This location usually contains a lot of high value loot."
+        ],
+        "video": "gxg4wNbfa8w",
+        "keyName": "Chemical Storage Warehouse Key"
+    },
+    {
+        "id": "ch-7-editorial-department",
+        "map": "al-mazrah",
+        "title": "Ch 7 Editorial Department",
+        "location": {
+            "x": 2418,
+            "y": 1222
+        },
+        "description": [
+            "Required for Flight Plans, Black Mous tier 5.",
+            "Taking the elevator in the south corner up to the top floor, you will see the door right across the hall."
+        ],
+        "video": "bVLUp0c7W5Q",
+        "tags": [
+            "mission"
+        ],
+        "keyName": "Ch 7 Editorial Department Key"
+    },
+    {
+        "id": "ch-7-secure-records-room",
+        "map": "al-mazrah",
+        "title": "Ch 7 Secure Records Room",
+        "location": {
+            "x": 2487,
+            "y": 1163
+        },
+        "description": [
+            "Entering the building from the south you will find the room right across the entrance."
+        ],
+        "video": "xieE3FJY838",
+        "keyName": "Ch 7 Secure Records Room Key"
+    },
+    {
+        "id": "clock-repair-tools",
+        "map": "al-mazrah",
+        "title": "Clock Repair Tools",
+        "location": {
+            "x": 2186,
+            "y": 3046
+        },
+        "description": [
+            "The clock tower is easy to spot from a distance. Taking the ladder inside the building takes you to the roof. From the roof you can climb another ladder to the top of the tower where the toolbox is located.",
+            "The key can be acquired by completing Custom Hardware, Black Mous tier 2"
+        ],
+        "video": "wUa_RhgDWoo",
+        "keyName": "Clock Repair Tools Key"
+    },
+    {
+        "id": "control-tower",
+        "map": "al-mazrah",
+        "title": "Control Tower",
+        "location": {
+            "x": 2840,
+            "y": 3134
+        },
+        "description": [
+            "Required for Flight Plans, Black Mous tier 5.",
+            "The Control Tower fortress is marked by a large icon on the map. There are many entrances to get into the building, and plenty of angry airplane enthousiasts to fight around the area.",
+            "The key can be acquired by completing Contractual Obligations, Legion tier 5"
+        ],
+        "video": "s71OBNU5YL4",
+        "tags": [
+            "fortress",
+            "mission"
+        ],
+        "keyName": "Control Tower Key"
+    },
+    {
+        "id": "crane-control-room",
+        "map": "al-mazrah",
+        "title": "Crane Control Room",
+        "location": {
+            "x": 792,
+            "y": 1643
+        },
+        "description": [
+            "Required for Unearthed, White Lotus tier 3.",
+            "The Crane Control Room is easy to spot as it's directly underneath a large crane at the Al-Safwa Quarry.",
+            "The key can be acquired by completing Quick and Dirty, Black Mous tier 3"
+        ],
+        "video": "0cHsa4jf3GQ",
+        "tags": [
+            "mission"
+        ],
+        "keyName": "Crane Control Room Key"
+    },
+    {
+        "id": "deckhands-toolbox",
+        "map": "al-mazrah",
+        "title": "Deckhand's Toolbox",
+        "location": {
+            "x": 1825,
+            "y": 3562
+        },
+        "description": [
+            "Getting to the toolbox requires swimming underneath the shipwreck. Dive from the northwest side through the underwater doorway next to the staircase, into the shipwreck. Then take the first left into the main corridor and keep swimming forward until you get to the toolbox.",
+            "The key can be acquired by completing One Man Army, Black Mous tier 1"
+        ],
+        "video": "EKw3PoOo99c",
+        "keyName": "Deckhand's Toolbox Key"
+    },
+    {
+        "id": "downtown-post-office",
+        "map": "al-mazrah",
+        "title": "Downtown Post Office",
+        "location": {
+            "x": 2342,
+            "y": 1733
+        },
+        "description": [
+            "Required for Clean House, White Lotus tier 5.",
+            "The Downtown Post Office is a fortress in Al-Mazrah City, as such it is marked by a large icon on the map. The building has lots of easy to find entrances. Enter to find out about the meaning of going postal."
+        ],
+        "video": "VTJJAmnRdnE",
+        "tags": [
+            "fortress",
+            "mission"
+        ],
+        "keyName": "Downtown Post Office Key"
+    },
+    {
+        "id": "downtown-post-office-secure-room",
+        "map": "al-mazrah",
+        "title": "Downtown Post Office Secure Room",
+        "location": {
+            "x": 2325,
+            "y": 1730
+        },
+        "description": [
+            "The Secure Room can be found on the top floor of the Downtown Post Office. As this location is a fortress, a separate key is needed to get to the door. The room will be the first door to your left exiting the southwest staircase."
+        ],
+        "video": "QmiU_Lpus0I",
+        "keyName": "Downtown Post Office Secure Room Key"
+    },
+    {
+        "id": "enfer-back-room",
+        "map": "al-mazrah",
+        "title": "Enfer Back Room",
+        "location": {
+            "x": 1421,
+            "y": 2634
+        },
+        "description": [
+            "<s>Noclip out of bounds to access the enfer backrooms.</s>",
+            "The Enfer Back Room is located inside the Sa'id City mall. Entrance into the mall through the roof is possible to bypass the extra key requirement."
+        ],
+        "video": "f5qbD-snl38",
+        "keyName": "Enfer Back Room Key"
+    },
+    {
+        "id": "far-south-eastern-storage-closet",
+        "map": "al-mazrah",
+        "title": "Far South Eastern Storage Closet",
+        "location": {
+            "x": 3101,
+            "y": 2294
+        },
+        "description": [
+            "On the ground floor of the north side of the marked building you will find the Far South Eastern Storage Closet."
+        ],
+        "video": "yz-PSu27vYs",
+        "keyName": "Far South Eastern Storage Closet Key"
+    },
+    {
+        "id": "groundskeepers-building",
+        "map": "al-mazrah",
+        "title": "Groundskeepers Building",
+        "location": {
+            "x": 1952,
+            "y": 3266
+        },
+        "description": [
+            "Small building in the west of Sarrif Bay. Entry from the far side of the road."
+        ],
+        "video": "mjYdGa0glRQ",
+        "keyName": "Groundskeepers Building Key"
+    },
+    {
+        "id": "hafid-ship-bridge-cache",
+        "map": "al-mazrah",
+        "title": "Hafid Ship Bridge Cache",
+        "location": {
+            "x": 585,
+            "y": 2575
+        },
+        "description": [
+            "Climb up three staircases from the deck of the Hafid Ship to get to the fourth floor overlooking the deck. You will find the toolbox on your right."
+        ],
+        "video": "VvQP9wCP7Zw",
+        "keyName": "Hafid Ship Bridge Cache Key"
+    },
+    {
+        "id": "hafid-ship-first-mate-cache",
+        "map": "al-mazrah",
+        "title": "Hafid Ship First Mate Cache",
+        "location": {
+            "x": 580,
+            "y": 2574
+        },
+        "description": [
+            "The Hafid Ship First Mate Cache can be found in the small room in the back of the third floor of the Hafid Ship."
+        ],
+        "video": "V20sBkaZ-hg",
+        "keyName": "Hafid Ship First Mate Cache Key"
+    },
+    {
+        "id": "zarqwa-hydro-island-computer",
+        "map": "al-mazrah",
+        "title": "Zarqwa Hydro Island Computer",
+        "location": {
+            "x": 1845,
+            "y": 1876
+        },
+        "description": [
+            "This computer can be found in the red power house in the center of the hydroelectric facility. The computer is located in the small room on the ground floor.",
+            "The key can be acquired by completing Intelligence Control, Legion tier 4"
+        ],
+        "video": "-g5M4VcuLtE",
+        "keyName": "Zarqwa Hydro Island Computer Key"
+    },
+    {
+        "id": "west-zarqwa-safe-house",
+        "map": "al-mazrah",
+        "title": "West Zarqwa Safe House",
+        "location": {
+            "x": 1736,
+            "y": 1923
+        },
+        "description": [
+            "The locked door is on the second floor of the marked building."
+        ],
+        "video": "c6kZZpUnSF8",
+        "keyName": "West Zarqwa Safe House Key"
+    },
+    {
+        "id": "ihtaq-warehouse-supply-room",
+        "map": "al-mazrah",
+        "title": "IHTAQ Warehouse Supply Room",
+        "location": {
+            "x": 946,
+            "y": 2556
+        },
+        "description": [
+            "This is a small room upstairs in one of the offices in the south side of the large hangar in Hafid Port."
+        ],
+        "video": "Txook-IBuXo",
+        "keyName": "IHTAQ Warehouse Supply Room Key"
+    },
+    {
+        "id": "kushaak-construction-warehouse",
+        "map": "al-mazrah",
+        "title": "Kushaak Construction Warehouse",
+        "location": {
+            "x": 826,
+            "y": 1735
+        },
+        "description": [
+            "Required for Clean House, White Lotus tier 5.",
+            "The Kushaak Construction Warehouse is a fortress right in the middle of Al-Safwa Quarry. Don't forget your hardhat.",
+            "The key can be acquired by completing Convenience, White Lotus tier 1 or Unearthed, White Lotus tier 3"
+        ],
+        "video": "cggyMST0Hb8",
+        "tags": [
+            "fortress",
+            "mission"
+        ],
+        "keyName": "Kushaak Construction Warehouse Key"
+    },
+    {
+        "id": "longshoremans-duffel-bag",
+        "map": "al-mazrah",
+        "title": "Longshoreman's Duffel Bag",
+        "location": {
+            "x": 675,
+            "y": 2700
+        },
+        "description": [
+            "The Duffel Bag is located at the very end of the large (broken) pier in Hafid Port pointing south."
+        ],
+        "video": "ycwVZcG2LsI",
+        "keyName": "Longshoreman's Duffel Bag Key"
+    },
+    {
+        "id": "mawizeh-cell-shop",
+        "map": "al-mazrah",
+        "title": "Mawizeh Cell Shop",
+        "location": {
+            "x": 2263,
+            "y": 2049
+        },
+        "description": [
+            "Near the Mawizeh Mashlands, following the large bridge going south are a few small buildings. The cell shop is located right next to the road heading northwest."
+        ],
+        "video": "QkKEd-BgRFU",
+        "tags": [
+            "fortress"
+        ],
+        "keyName": "Mawizeh Cell Shop Key"
+    },
+    {
+        "id": "mawizeh-power-control-substation",
+        "map": "al-mazrah",
+        "title": "Mawizeh Power Control Substation",
+        "location": {
+            "x": 2522,
+            "y": 1943
+        },
+        "description": [
+            "The Mawizeh Power Control Substation is located across the Mashlands from the Mawizeh Resort. It's the small building next to the large white watertower."
+        ],
+        "video": "BNxTrs1CjEE",
+        "keyName": "Mawizeh Power Control Substation Key"
+    },
+    {
+        "id": "mawizeh-resort-bungalow-room",
+        "map": "al-mazrah",
+        "title": "Mawizeh Resort Bungalow Room",
+        "location": {
+            "x": 2411,
+            "y": 2081
+        },
+        "description": [
+            "The Bungalow Room is the third bungalow counting from the west of the large Mawizeh Marshland island."
+        ],
+        "video": "bbGPxQVEu9g",
+        "keyName": "Mawizeh Resort Bungalow Room Key"
+    },
+    {
+        "id": "mawizeh-resort-main-building",
+        "map": "al-mazrah",
+        "title": "Mawizeh Resort Main Building",
+        "location": {
+            "x": 2525,
+            "y": 2095
+        },
+        "description": [
+            "The Mawizeh Resort Main Building is marked on the map with a large icon, because it's a fortress. The complex consists of both of the large buildings on the Mawizeh Marshlands island.",
+            "The key can be acquired by completing Harmful Waves, Black Mous tier 4"
+        ],
+        "video": "JMAehvF_uIM",
+        "tags": [
+            "fortress"
+        ],
+        "keyName": "Mawizeh Resort Main Building Key"
+    },
+    {
+        "id": "mawizah-slum-discarded-cache",
+        "map": "al-mazrah",
+        "title": "Mawizah Slum Discarded Cache",
+        "location": {
+            "x": 2325,
+            "y": 2035
+        },
+        "description": [
+            "The Discarded Cache is located underneath the large bridge on the south side of the river, where it's laying on the ground next to a burned out car."
+        ],
+        "video": "j3qq7WN7-HA",
+        "keyName": "Mawizah Slum Discarded Cache Key"
+    },
+    {
+        "id": "mountaintop-spotter-shack",
+        "map": "al-mazrah",
+        "title": "Mountaintop Spotter Shack",
+        "location": {
+            "x": 2642,
+            "y": 2455
+        },
+        "description": [
+            "The Mountaintop, as the name implies, is located on top of the mountain at Al Sharim Pass. It's a small building at the northeastern edge of the mountain."
+        ],
+        "video": "5zhNOSGTuiY",
+        "keyName": "Mountaintop Spotter Shack Key"
+    },
+    {
+        "id": "north-canals-info-booth",
+        "map": "al-mazrah",
+        "title": "North Canals Info Booth",
+        "location": {
+            "x": 2299,
+            "y": 1469
+        },
+        "description": [
+            "Along the canal running through Al-Mazrah City are multiple booths to the west. This booth (as the name implies) is the northernmost one."
+        ],
+        "video": "n8TU8OBDeHk",
+        "keyName": "North Canals Info Booth Key"
+    },
+    {
+        "id": "old-lighthouse-citadel-room",
+        "map": "al-mazrah",
+        "title": "Old Lighthouse Citadel Room",
+        "location": {
+            "x": 2108,
+            "y": 3359
+        },
+        "description": [
+            "At the end of the island at Sarrif Bay is a small lighthouse. Going inside you will quickly find the door right in front of you."
+        ],
+        "video": "MdYb5uV5aw0",
+        "keyName": "Old Lighthouse Citadel Room Key"
+    },
+    {
+        "id": "omars-auto-repair",
+        "map": "al-mazrah",
+        "title": "Omar's Auto Repair",
+        "location": {
+            "x": 1251,
+            "y": 3243
+        },
+        "description": [
+            "Omar's Auto Repair is located inside of Sawah Village. Note that the village usually contains a lot of AI. The door is located in the back of the car garage building."
+        ],
+        "video": "Zf2Z614ZEtw",
+        "keyName": "Omar's Auto Repair Key"
+    },
+    {
+        "id": "police-academy",
+        "map": "al-mazrah",
+        "title": "Police Academy",
+        "location": {
+            "x": 2848,
+            "y": 1605
+        },
+        "description": [
+            "The Police Academy is marked by a large icon in the east of the map. As this is a fortress it contains a lot of enemies. Note that the Academy has a lot of doors on both the inside and outside, recommended to prioritize opening the inside doors.",
+            "The key can be acquired by completing The Golden Rule, Black Mous tier 4"
+        ],
+        "video": "ce19mYClMao",
+        "tags": [
+            "fortress"
+        ],
+        "keyName": "Police Academy Key"
+    },
+    {
+        "id": "police-armory-f2",
+        "map": "al-mazrah",
+        "title": "Police Armory [F2]",
+        "keyName": "Police Armory Key",
+        "location": {
+            "x": 2204,
+            "y": 1233
+        },
+        "description": [
+            "One of three locations for the Police Armory Key.",
+            "The Police Armory is located inside the police building, up the stairs on the left. This room itself contains 8 Police Lockers which require a Police Locker Master Key to open, this key has a chance to spawn inside the room."
+        ],
+        "video": "32_9NKyg44g"
+    },
+    {
+        "id": "police-locker-f2",
+        "map": "al-mazrah",
+        "title": "Police Locker [F2]",
+        "keyName": "Police Locker Master Key",
+        "location": {
+            "x": 2194,
+            "y": 1227
+        },
+        "description": [
+            "One of three locations for the Police Locker Master Key.",
+            "Inside of the Police Armory. Key required for the room to get to the lockers."
+        ],
+        "video": "s0nVFQfdRUw"
+    },
+    {
+        "id": "police-armory-g5",
+        "map": "al-mazrah",
+        "title": "Police Armory [G5]",
+        "keyName": "Police Armory Key",
+        "location": {
+            "x": 2824,
+            "y": 2086
+        },
+        "description": [
+            "One of three locations for the Police Armory Key.",
+            "The Police Armory is located inside the police building, up the stairs on the left. This room itself contains 8 Police Lockers which require a Police Locker Master Key to open, this key has a chance to spawn inside the room."
+        ],
+        "video": "tNtirUj6nIQ"
+    },
+    {
+        "id": "police-locker-g5",
+        "map": "al-mazrah",
+        "title": "Police Locker [G5]",
+        "keyName": "Police Locker Master Key",
+        "location": {
+            "x": 2818,
+            "y": 2099
+        },
+        "description": [
+            "One of four locations for the Police Locker Master Key.",
+            "Inside of the Police Armory. Key required for the room to get to the lockers."
+        ],
+        "video": "Sl1bFtO9yFI"
+    },
+    {
+        "id": "police-armory-d6",
+        "map": "al-mazrah",
+        "title": "Police Armory [D6]",
+        "keyName": "Police Armory Key",
+        "location": {
+            "x": 1477,
+            "y": 2693
+        },
+        "description": [
+            "One of four locations for the Police Armory Key.",
+            "The Police Armory is located inside the police building, up the stairs on the left. This room itself contains 8 Police Lockers which require a Police Locker Master Key to open, this key has a chance to spawn inside the room."
+        ],
+        "video": "bx3v2ZeghSc"
+    },
+    {
+        "id": "police-locker-d6",
+        "map": "al-mazrah",
+        "title": "Police Locker [D6]",
+        "keyName": "Police Locker Master Key",
+        "location": {
+            "x": 1491,
+            "y": 2693
+        },
+        "description": [
+            "One of four locations for the Police Locker Master Key.",
+            "Inside of the Police Armory. Key required for the room to get to the lockers."
+        ],
+        "video": "72bSwO7tkG0"
+    },
+    {
+        "id": "police-academy-private-locker-room",
+        "map": "al-mazrah",
+        "title": "Police Academy Private Locker Room",
+        "location": {
+            "x": 2885,
+            "y": 1708
+        },
+        "description": [
+            "The Police Academy Private Locker Room is located downstairs inside the building to the south of the Police Academy parking lot."
+        ],
+        "video": "CA5ibOFOCCg",
+        "keyName": "Police Academy Private Locker Room Key"
+    },
+    {
+        "id": "police-locker-g4",
+        "map": "al-mazrah",
+        "title": "Police Locker [G4]",
+        "keyName": "Police Locker Master Key",
+        "location": {
+            "x": 2885,
+            "y": 1712
+        },
+        "description": [
+            "One of four locations for the Police Locker Master Key.",
+            "Inside of the Police Academy Private Locker Room. Key required for the room to get to the lockers."
+        ]
+    },
+    {
+        "id": "police-academy-server-admin",
+        "map": "al-mazrah",
+        "title": "Police Academy Server Admin",
+        "location": {
+            "x": 2900,
+            "y": 1699
+        },
+        "description": [
+            "To the south of the large parking lot at the Police Academy is a smaller building. The room is on the ground floor in the north of the building."
+        ],
+        "video": "OrK3UZ-b3wk",
+        "keyName": "Police Academy Server Admin Key"
+    },
+    {
+        "id": "power-substation-toolbox",
+        "map": "al-mazrah",
+        "title": "Power Substation Toolbox",
+        "location": {
+            "x": 1475,
+            "y": 3127
+        },
+        "description": [
+            "To the south of the Al Samman Cemetery is a small Power Substation. Climbing over the fence, the toolbox is located right in the middle of the station."
+        ],
+        "video": "XFnwZRN84ng",
+        "keyName": "Power Substation Toolbox Key"
+    },
+    {
+        "id": "quarry-workers-lost-toolbox",
+        "map": "al-mazrah",
+        "title": "Quarry Worker's Lost Toolbox",
+        "location": {
+            "x": 768,
+            "y": 1703
+        },
+        "description": [
+            "The Lost Toolbox is located underwater in the north of the Al-Safwa Quarry.",
+            "The key can be acquired by completing Bombing Run, White Lotus tier 2"
+        ],
+        "video": "_alliZFxcZI",
+        "keyName": "Quarry Worker's Lost Toolbox Key"
+    },
+    {
+        "id": "special-ops-relay-station",
+        "map": "al-mazrah",
+        "title": "Special Ops Relay Station",
+        "location": {
+            "x": 809,
+            "y": 2720
+        },
+        "description": [
+            "Next to the T-shaped dock on the south of Hafid Port is a small building. This is the Special Ops Relay Station.",
+            "The key can be acquired by completing Battle Rivals, Legion tier 5"
+        ],
+        "video": "bYneoNHg5bU",
+        "keyName": "Special Ops Relay Station Key"
+    },
+    {
+        "id": "retro-runway-fridge",
+        "map": "al-mazrah",
+        "title": "Retro Runway Fridge",
+        "location": {
+            "x": 3262,
+            "y": 2902
+        },
+        "description": [
+            "The Retro Runway Fridge is a medium sized white box commonly used for cooling, located inside the crater at the eastern end of the Al Malik Airport runway."
+        ],
+        "video": "0DvogRQ46h4",
+        "keyName": "Retro Runway Fridge Key"
+    },
+    {
+        "id": "rohan-control-room-locker",
+        "map": "al-mazrah",
+        "title": "Rohan Control Room Locker",
+        "location": {
+            "x": 1404,
+            "y": 1431
+        },
+        "description": [
+            "The Rohan Control Room Locker is a locker underground in the Rohan Oil Tunnels. There are multiple entrances to these tunnels around the facility."
+        ],
+        "video": "SIJvnqZF5tA",
+        "keyName": "Rohan Control Room Locker Key"
+    },
+    {
+        "id": "rohan-oil-north-guard-shack",
+        "map": "al-mazrah",
+        "title": "Rohan Oil North Guard Shack",
+        "location": {
+            "x": 1470,
+            "y": 1263
+        },
+        "description": [
+            "This small guard shack is located next to the south gate at Rohan Oil."
+        ],
+        "video": "Dc4zL9fBv4U",
+        "keyName": "Rohan Oil North Guard Shack Key"
+    },
+    {
+        "id": "rohan-oil-south-guard-shack",
+        "map": "al-mazrah",
+        "title": "Rohan Oil South Guard Shack",
+        "location": {
+            "x": 1650,
+            "y": 1582
+        },
+        "description": [
+            "This small guard shack is located next to the south gate at Rohan Oil."
+        ],
+        "video": "JEbusK_AiFE",
+        "keyName": "Rohan Oil South Guard Shack Key"
+    },
+    {
+        "id": "rohan-utility-tunnel-locker",
+        "map": "al-mazrah",
+        "title": "Rohan Utility Tunnel Locker",
+        "location": {
+            "x": 1463,
+            "y": 1324
+        },
+        "description": [
+            "This locker is underground, with the entrance in the marked building just north of the train tracks."
+        ],
+        "video": "L4HcdGP1q4A",
+        "keyName": "Rohan Utility Tunnel Locker Key"
+    },
+    {
+        "id": "said-mall-security-room",
+        "map": "al-mazrah",
+        "title": "Sa'id Mall Security Room",
+        "location": {
+            "x": 1374,
+            "y": 2625
+        },
+        "description": [
+            "The Security Room is inside of the Sa'id Mall in the east of the map. The room itself is located upstairs on the right side of the mall at the end of the hallway."
+        ],
+        "video": "myYuLlmhSm8",
+        "keyName": "Sa'id Mall Security Room Key"
+    },
+    {
+        "id": "sattiq-guest-quarters",
+        "map": "al-mazrah",
+        "title": "Sattiq Guest Quarters",
+        "location": {
+            "x": 1383,
+            "y": 2042
+        },
+        "description": [
+            "This building is a little hard to find, as it's not marked on the map. It's a small door above ground in the side of a hill."
+        ],
+        "video": "fn9ytgMuMP8",
+        "keyName": "Sattiq Guest Quarters Key"
+    },
+    {
+        "id": "sattiq-poppy-farm-house",
+        "map": "al-mazrah",
+        "title": "Sattiq Poppy Farm House",
+        "location": {
+            "x": 1218,
+            "y": 1968
+        },
+        "description": [
+            "Required for Espionage, Black Mous tier 2.",
+            "The Sattiq Poppy Farm House is a small building in the north of the circular road at the Sattiq mountains. Note that you need to enter the room twice to complete the mission.",
+            "The key can be acquired by completing Robin Hoofd, White Lotus tier 3"
+        ],
+        "video": "uaw5eT_RAC4",
+        "tags": [
+            "mission"
+        ],
+        "keyName": "Sattiq Poppy Farm House Key"
+    },
+    {
+        "id": "sawah-crypto-mining-farm-key",
+        "map": "al-mazrah",
+        "title": "Sawah Crypto Mining Farm Key",
+        "location": {
+            "x": 1276,
+            "y": 3321
+        },
+        "description": [
+            "To the west of the Sawah Village square is a large building. Going up the large stairs and taking an immediate right in the hallway leaves you at the door."
+        ],
+        "video": "By_6_Y4WAWs",
+        "keyName": "Sawah Crypto Mining Farm Key Key"
+    },
+    {
+        "id": "sawah-hotel-room-302",
+        "map": "al-mazrah",
+        "title": "Sawah Hotel Room 302",
+        "location": {
+            "x": 1370,
+            "y": 3295
+        },
+        "description": [
+            "Required for Good Fortune, Legion Tier 3.",
+            "The door is located on the top floor of the hotel, right next to Sawah Hotel Room 303. Entry without key is possible through the skylights when the hotel is not a fortress."
+        ],
+        "video": "g6UGMlEvhMM",
+        "tags": [
+            "mission"
+        ],
+        "keyName": "Sawah Hotel Room 302 Key"
+    },
+    {
+        "id": "sawah-hotel-room-303",
+        "map": "al-mazrah",
+        "title": "Sawah Hotel Room 303",
+        "location": {
+            "x": 1374,
+            "y": 3314
+        },
+        "description": [
+            "The door is located on the top floor of the hotel, right next to Sawah Hotel Room 302. Entry without key is possible through the skylights when the hotel is not a fortress.",
+            "The key can be acquired by completing Cover No Concealment, White Lotus tier 4"
+        ],
+        "video": "g6UGMlEvhMM",
+        "keyName": "Sawah Hotel Room 303 Key"
+    },
+    {
+        "id": "scientists-locker",
+        "map": "al-mazrah",
+        "title": "Scientist's Locker",
+        "location": {
+            "x": 2100,
+            "y": 2350
+        },
+        "description": [
+            "The Scientist's Locker is located at the Zaya Observatory, inside the bunker to the north of the complex. Following the tunnel east you will find the locker on your right about halfway through."
+        ],
+        "video": "iPEj1jqQMWM",
+        "keyName": "Scientist's Locker Key"
+    },
+    {
+        "id": "south-al-bagra-top-floor-bedroom",
+        "map": "al-mazrah",
+        "title": "South Al Bagra Top Floor Bedroom",
+        "location": {
+            "x": 2650,
+            "y": 3463
+        },
+        "description": [
+            "Unsurprisingly found on the top floor of the building, in the southwest corner of the floor."
+        ],
+        "video": "s3rW3NX1HqM",
+        "keyName": "South Al Bagra Top Floor Bedroom Key"
+    },
+    {
+        "id": "south-bank-apartment",
+        "map": "al-mazrah",
+        "title": "South Bank Apartment",
+        "location": {
+            "x": 1346,
+            "y": 2868
+        },
+        "description": [
+            "On the other side of the river under Sa'id City is a small town. Entering the marked building from the roof, you will immediately find the door on the left."
+        ],
+        "video": "Jn3ihx9ArLE",
+        "keyName": "South Bank Apartment Key"
+    },
+    {
+        "id": "south-zaya-scientist-apartment",
+        "map": "al-mazrah",
+        "title": "South Zaya Scientist Apartment",
+        "location": {
+            "x": 2010,
+            "y": 2654
+        },
+        "description": [
+            "The Scientist Apartment is located in the south of the Zaya Observatory. Going up the central stairs in the marked building and taking a left there is a single door to the apartment."
+        ],
+        "video": "ISyC9NQPJEg",
+        "keyName": "South Zaya Scientist Apartment Key"
+    },
+    {
+        "id": "special-forces-dead-drop",
+        "map": "al-mazrah",
+        "title": "Special Forces Dead Drop",
+        "location": {
+            "x": 2595,
+            "y": 1090
+        },
+        "description": [
+            "The Special Forces Dead Drop is a lockbox located on the concrete platform underneath the large tower in the northeast of the map."
+        ],
+        "video": "bcImnQY-ks8",
+        "keyName": "Special Forces Dead Drop Key"
+    },
+    {
+        "id": "taraq-river-supply-shack",
+        "map": "al-mazrah",
+        "title": "Taraq River Supply Shack",
+        "location": {
+            "x": 2030,
+            "y": 910
+        },
+        "description": [
+            "The Taraq River Supply Shack is located next to the river to the east of Taraq Village. The small building is not outlined on the map but easy to find as it stands out next to the river."
+        ],
+        "video": "VPZIVFIWvyk",
+        "keyName": "Taraq River Supply Shack Key"
+    },
+    {
+        "id": "taraq-smugglers-office",
+        "map": "al-mazrah",
+        "title": "Taraq Smuggler's Office",
+        "location": {
+            "x": 1981,
+            "y": 900
+        },
+        "description": [
+            "Crossing the bridge heading to Taraq Village are three buildings on the right of the road. The Smuggler's Office is located inside the middle building on the ground floor."
+        ],
+        "video": "WUUxiX0pMpQ",
+        "keyName": "Taraq Smuggler's Office Key"
+    },
+    {
+        "id": "taraq-tunnels-workbench-cache",
+        "map": "al-mazrah",
+        "title": "Taraq Tunnels Workbench Cache",
+        "location": {
+            "x": 1900,
+            "y": 1030
+        },
+        "description": [
+            "To the west of the building next to the marked location is a hole in the ground. Entering this hole and following the tunnel you will find the locked toolbox next to a workbench."
+        ],
+        "video": "n2N8ZB6_YFk",
+        "keyName": "Taraq Tunnels Workbench Cache Key"
+    },
+    {
+        "id": "tower-conference-room",
+        "map": "al-mazrah",
+        "title": "Tower Conference Room",
+        "location": {
+            "x": 2846,
+            "y": 3112
+        },
+        "description": [
+            "The Tower Conference Room is located inside the Control Tower fortress, so you need the respective key to get to this room. The room itself is in the northeast corner of the first floor, just to the right of the top of the escalators."
+        ],
+        "video": "p0-HK4yHmZg",
+        "keyName": "Tower Conference Room Key"
+    },
+    {
+        "id": "travellers-luggage",
+        "map": "al-mazrah",
+        "title": "Traveller's Luggage",
+        "location": {
+            "x": 2836,
+            "y": 2925
+        },
+        "description": [
+            "The Traveller's Luggage is a duffel bag located in the small waiting area overlooking the Al Malik Airport.",
+            "The key can be acquired by completing Dead Drop, Legion tier 3"
+        ],
+        "video": "q8DSYEJS6u4",
+        "keyName": "Traveller's Luggage Key"
+    },
+    {
+        "id": "under-freeway-east-warehouse",
+        "map": "al-mazrah",
+        "title": "Under Freeway East Warehouse",
+        "location": {
+            "x": 2866,
+            "y": 2000
+        },
+        "description": [
+            "The marked building is located right next to the road going under the highway. The door is on the outside of this building."
+        ],
+        "video": "mIXoneswWKQ",
+        "keyName": "Under Freeway East Warehouse Key"
+    },
+    {
+        "id": "us-embassy",
+        "map": "al-mazrah",
+        "title": "U.S. Embassy",
+        "location": {
+            "x": 2475,
+            "y": 1315
+        },
+        "description": [
+            "Required for Clean House, White Lotus tier 5.",
+            "The U.S. Embassy is a large building marked by an icon on the map inside of Al-Mazrah City. There are multiple entrances around the building. Expect a lot of enemies as this is a fortress.",
+            "The key can be acquired by completing The Client Is King, Black Mous tier 5"
+        ],
+        "video": "4fl5saUVHKA",
+        "tags": [
+            "fortress",
+            "mission"
+        ],
+        "keyName": "U.S. Embassy Key"
+    },
+    {
+        "id": "weapons-repair-kit",
+        "map": "al-mazrah",
+        "title": "Weapons Repair Kit",
+        "location": {
+            "x": 1254,
+            "y": 2103
+        },
+        "description": [
+            "The Weapons Repair Kit is located inside of a cave at the Sattiq Cave Complex. The cave entrance is to the north of the marker. Heading inside and going right you will find the toolbox on a table."
+        ],
+        "video": "IH39322Yxtg",
+        "keyName": "Weapons Repair Kit Key"
+    },
+    {
+        "id": "yum-yum-burger-back-room",
+        "map": "al-mazrah",
+        "title": "Yum Yum Burger Back Room",
+        "location": {
+            "x": 2549,
+            "y": 1418
+        },
+        "description": [
+            "You will find the Back Room door right next to the staircase on the ground floor inside of the Yum Yum Burger building."
+        ],
+        "video": "K3xIASt5KwY",
+        "keyName": "Yum Yum Burger Back Room Key"
+    },
+    {
+        "id": "zarqwa-market-storage",
+        "map": "al-mazrah",
+        "title": "Zarqwa Market Storage",
+        "location": {
+            "x": 1875,
+            "y": 1918
+        },
+        "description": [
+            "The Market Storage is a small building next to the river running through the Zarqwa village. It's located inside of a small covered market area."
+        ],
+        "video": "90Rh-WTItAI",
+        "keyName": "Zarqwa Market Storage Key"
+    },
+    {
+        "id": "zarqwa-red-village-appartment",
+        "map": "al-mazrah",
+        "title": "Zarqwa Red Village Appartment",
+        "location": {
+            "x": 1948,
+            "y": 1902
+        },
+        "description": [
+            "The Zarqwa Red Village Apartment is a small apartment upstairs in the marked building in Zarqwa village."
+        ],
+        "video": "ircnrYJKUfM",
+        "keyName": "Zarqwa Red Village Appartment Key"
+    },
+    {
+        "id": "zaya-radar-dome",
+        "map": "al-mazrah",
+        "title": "Zaya Radar Dome",
+        "location": {
+            "x": 2035,
+            "y": 2371
+        },
+        "description": [
+            "The Zaya Radar Dome key leads to one of the large Radar buildings in Zaya Observatory. This one is located to the west of the bunker, with an easy to spot entrace at the ground floor coming from the east."
+        ],
+        "video": "LAaajr_4_QQ",
+        "keyName": "Zaya Radar Dome Key"
+    },
+    {
+        "id": "sunken-ship-thiefs-cache",
+        "map": "al-mazrah",
+        "title": "Sunken Ship Thief's Cache",
+        "location": {
+            "x": 1791,
+            "y": 3580
+        },
+        "description": [
+            "The Thief's Cache is inside an underwater part of the sunken ship. Entering the hull from the front of the ship you will find the duffel bag on your immediate right. Contains the $25,000 Golden Skull of Al Bagra Minor."
+        ],
+        "video": "Rx4L86BCsX8",
+        "keyName": "Sunken Ship Thief's Cache Key"
+    },
+    {
+        "id": "control-room",
+        "map": "al-mazrah",
+        "title": "Control Room",
+        "description": [
+            "If you find the location for this key, please contact me in Discord and I will add it to the map."
+        ],
+        "keyName": "Control Room Key"
+    },
+    {
+        "id": "fort-barrack",
+        "map": "al-mazrah",
+        "title": "Fort Barrack",
+        "description": [
+            "Might work for Al Bagra Barrack location.",
+            "If you find the location for this key, please contact me in Discord and I will add it to the map."
+        ],
+        "keyName": "Fort Barrack Key"
+    },
+    {
+        "id": "storage",
+        "map": "al-mazrah",
+        "title": "Storage",
+        "description": [
+            "If you find the location for this key, please contact me in Discord and I will add it to the map."
+        ],
+        "keyName": "Storage Key"
+    },
+    {
+        "id": "foremans-office",
+        "map": "al-mazrah",
+        "title": "Foreman's Office",
+        "description": [
+            "If you find the location for this key, please contact me in Discord and I will add it to the map."
+        ],
+        "keyName": "Foreman's Office Key"
+    },
+    {
+        "id": "stronghold-interior",
+        "map": "al-mazrah",
+        "title": "Stronghold Interior",
+        "description": [
+            "Sometimes disappears from the inventory on extraction.",
+            "If you find the location for this key, please contact me in Discord and I will add it to the map."
+        ],
+        "keyName": "Stronghold Interior Key"
+    },
+    {
+        "id": "ashika-island-test-location",
+        "map": "ashika-island",
+        "title": "Ashika Island Test Location",
+        "location": {
+            "x": 234,
+            "y": 3453
+        },
+        "description": [
+            "Easy to find, this is the office room inside of the grocery store."
+        ],
+        "video": "xfyHNVdPqQs",
+        "keyName": "Ashika Island Test Location Key"
+    }
+]
