+++ src/data/maps.js
@@ -1,8 +1,13 @@
 import { data as alMazrahData } from "./al-mazrah/data";
+import { data as ashikaIslandData } from "./ashika-island/data";
 
 export const maps = {
 	"al-mazrah": {
 		title: "Al Mazrah",
 		data: alMazrahData,
+	},
+	"ashika-island": {
+		title: "Ashika Island",
+		data: ashikaIslandData,
 	},
 };
