--- src/routes/+layout.svelte
@@ -1,6 +1,0 @@
-<script>
-	import "leaflet/dist/leaflet.css";
-	import "../map.css";
-</script>
-
-<slot />
